Refactor plugin, hooks, settings, code namespaced with psr4
