+++ .gitignore
@@ -1,0 +1,9 @@
+.DS_Store
+/node_modules/
+/npm-debug.log
+/vendor/
+/dist/
+/tests/logs/
+/wordpress/
+build
+.phpunit.result.cache
